+++ STAGGI-Budget-API/Services/Interfaces/ICategoryService.cs
@@ -9,8 +9,8 @@ namespace STAGGI_Budget_API.Services.Interfaces
     {
         public Result<List<CategoryDTO>> GetAll();
         public Result<List<CategoryDTO>> GetByUserEmail(string email);
-        public Result<CategoryDTO> FindById(long id);
+        public Result<CategoryDTO> FindById(int id);
         public Result<string> CreateCategory(CategoryDTO categoryDTO, string email);
-        public Result<string> UpdateCategory(long Id, CategoryDTO categoryDTO, string email);
+        public Result<string> UpdateCategory(int Id, CategoryDTO categoryDTO, string email);
     }
 }
